Create missing skills chart
