+++ app/lib/getUserJobSkillsAndFrequency.ts
@@ -2,323 +2,8 @@
 
 import getCurrentUser from "./getCurrentUser";
 import prisma from "./db/prisma";
+import { extractSkillsFromDescription } from "./extractSkillsFromDescription";
 
-const skillKeywords = [
-  "Emotion",
-  "RxJS",
-  "Mailchimp",
-  "Relational Database",
-  "Nonrelational Database",
-  "Sequelize",
-  "Data Structures",
-  "Algorithms",
-  "MariaDB",
-  "SQLite",
-  "Django",
-  "Spring Boot",
-  "Spring",
-  "SEO",
-  "Ruby",
-  "Mongoose",
-  "Google Tag Manager",
-  "Google Analytics",
-  "Shopify",
-  "Stripe",
-  "PHP",
-  "jQuery",
-  "MySQL",
-  "Git",
-  "Github",
-  "Version Control",
-  "CI / CD",
-  "ElasticSearch",
-  "Azure",
-  "HTML",
-  "HTML5",
-  "CSS",
-  "CSS3",
-  "TailwindCSS",
-  "Tailwind",
-  "Nest",
-  "Nest.js",
-  "LESS",
-  "SASS",
-  "TypeScript",
-  "JavaScript",
-  "React",
-  "Angular",
-  "Redux",
-  "React Native",
-  "Next",
-  "Nextjs",
-  "Next.js",
-  "React.js",
-  "PostgreSQL",
-  "Microsoft SQL Server",
-  "MongoDB",
-  "Prisma",
-  "NoSQL",
-  "SQL",
-  "Webpack",
-  "Java",
-  "Objective-C",
-  "Go (Golang)",
-  "Golang",
-  "GraphQL",
-  "Apollo",
-  "Python",
-  "Flask",
-  "C ++",
-  "C#",
-  "Node",
-  "Express",
-  "Websockets",
-  "Web sockets",
-  "GraphQL Subscriptions",
-  ".NET",
-  "Vue",
-  "AWS",
-  "Firebase",
-  "Firebase Realtime Database",
-  "Firebase Cloud Firestore",
-  "Three.js",
-  "Figma",
-  "Expo",
-  "Docker",
-  "Kubernetes",
-  "Kubeflow",
-  "Cloudflare",
-  "Ruby on Rails",
-  "Storybook",
-  "Jest",
-  "Cypress",
-  "Mocha",
-  "Puppeteer",
-  "Chai",
-  "Jasmine",
-  "Karma",
-  "Enzyme",
-  "Selenium",
-  "Rust",
-  "Microsoft Azure",
-  "Hasura",
-  "Kotlin",
-  "Redis",
-  "Kafka",
-  "Sentry",
-  "Datadog",
-  "Pagerduty",
-  "OpenTelemetry",
-  "Terraform",
-  "Restful APIs",
-  "RestAPI",
-  "Rest API",
-  "OAuth",
-  "GCP",
-  "Google Cloud",
-  "Google Cloud Platform",
-  "Stable Diffusion",
-  "LLM",
-  "Swift",
-  "Cassandra",
-  "ScyllaDB",
-  "ErlangVM",
-  "WebGL",
-  "Mapbox",
-  "Odoo",
-  "Scala",
-  "Ruby",
-  "Laravel",
-  "Wordpress",
-  "SvelteJS",
-  "Postgres",
-  "DynamoDB",
-  "Kinesis",
-  "Twilio",
-  "Sendgrid",
-  "CMS",
-  "Content Management System",
-  "Headless CMS",
-  "Contentful",
-  "Prismic",
-  "Solana",
-  "Metaplex",
-  "tRPC",
-  "Pulumi",
-  "Data Visualization",
-  "D3",
-  "Plotly",
-  "Android",
-  "Pytorch",
-  "WebRTC",
-  "MobX",
-  "Jotai",
-  "Zustand",
-  "XState",
-  "Hasura",
-  "Preact",
-  "Solidjs",
-  "Langchain",
-  "LlamdaIndex",
-  "FastAPI",
-  "Machine Learning",
-  "ThreeJS",
-  "Blender",
-  "Ember",
-  "Lambda",
-  "Vercel",
-  "Netlify",
-  "Heroku",
-  "Digital Ocean",
-  "Rales",
-  "NLP",
-  "Natural Language Processing",
-  "Lucene",
-  "Xcode",
-  "React Query",
-  "Tanstack",
-  "useSWR",
-  "Chrome Extensions",
-  "Artifical Intelligence",
-  "Wordpress Plugin",
-  "LLMs",
-  "GraphQL Subscriptions",
-  "GraphQL Mutations",
-  "GraphQL Queries",
-  "Elixir",
-  "Phoenix",
-  "Apache",
-  "Jenkins",
-  "TDD",
-  "React Spring",
-  "Framer Motion",
-  "React Testing Library",
-  "Circle.ci",
-  "CircleCI",
-  "Github Actions",
-  "GitLab CI",
-  "GitlabCI",
-  "CloudFormation",
-  "Neo4j",
-  "Flowscript",
-  "Apache Airflow",
-  "Material-UI",
-  "MUI",
-  "Probe",
-  "Umami",
-  "FFmpeg",
-  ".NET",
-  "Amazon Web Services",
-  "Unity3D",
-  "TensorFlow",
-  "Theano",
-  "OOP",
-  "Firestore",
-  "Blockchain",
-  "Smart Contracts",
-  "Amazon S3",
-  "Celery",
-  "AWS SageMaker",
-  "A/B Testing",
-  "PySpark",
-  "OLAP SQL",
-  "K8s",
-  "WooCommerce",
-  "BigCommerce",
-  "Progressive Web App",
-  "Ethereum",
-  "Solidity",
-  "Power Platform",
-  "Microsoft 365",
-  "Dynamics 365",
-  "API Documentation",
-  "RabbitMQ",
-  "Symphony",
-  "VertX",
-  "JWT Tokens",
-  "Test Driven Development",
-  "Fargate",
-  "Stripe API",
-  "Jira",
-  "Trello",
-  "Fastify",
-  "S3",
-  "Svelte",
-  "CloudWatch",
-  "Kinesis",
-  "Elastic Search",
-  "Erlang",
-  "Ionic",
-  "Bigtable",
-  "Supabase",
-  "ASP.NET",
-  "Flux",
-  "Babel",
-  "Chart.js",
-  "Highcharts",
-  "Highcharts.js",
-  "BitBucket",
-  "VanillaJS",
-  "Vanilla JS",
-  "Vanilla JavaScript",
-  "Electron",
-  "HTMX",
-  "Middleware",
-  "Vitest",
-  "Automated testing",
-  "Flutter",
-  "NLTK",
-  "NumPY",
-  "SciPY",
-  "Keras",
-  "Unreal Engine",
-  "PowerBI",
-  "Linux",
-  "RTKQuery",
-  "AWS Amplify",
-  "Elastic Beanstalk",
-  "Amazon Redshift",
-  "Amazon SNS",
-  "Amazon SQS",
-  "Amazon AWS Glue",
-  "Data Lake",
-  "AWS EventBridge",
-  "Unit Testing",
-  "Unit Tests",
-  "Integration Testing",
-  "Integration Tests",
-  "Microservices",
-  "JUnit",
-  "TestNG",
-  "Relay",
-  "Haskell",
-  "Clojure",
-  "Perl",
-];
-
-const extractSkillsFromDescription = (description: string): string[] => {
-  // Convert the description to lowercase for case-insensitive comparison
-  const lowercaseDescription = description.toLowerCase();
-  // Filter the skillKeywords array based on certain conditions
-
-  const filteredSkills = skillKeywords.filter((skill) => {
-    // Convert each skill to lowercase for case-insensitive comparison
-    const lowercaseSkill = skill.toLowerCase();
-    // Check if the skill is 'chai' and the description contains certain irrelevant keywords related to blockchains
-    if (
-      lowercaseSkill === "chai" &&
-      (lowercaseDescription.includes("blockchains") ||
-        lowercaseDescription.includes("blockchain") ||
-        lowercaseDescription.includes("chain") ||
-        lowercaseDescription.includes("chains"))
-    ) {
-      return false;
-    }
-    return lowercaseDescription.includes(lowercaseSkill);
-  });
-
-  return filteredSkills.length > 0 ? filteredSkills : ["No skills available"];
-};
 
 const updateSkillFrequencyMap = (
   skillFrequencyMap: Map<string, number>,
